pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infrabot"
version = "0.1.0"
description = "AI-powered DevOps assistant for Ansible tasks and Linux server management"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["devops", "ansible", "ai", "automation", "infrastructure"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "crewai>=0.55.0",
    "chromadb>=1.0.20",
    "langchain-community>=0.3.29",
    "langchain-ollama>=0.3.29",
    "ansible>=9.0.0",
    "ollama>=0.5.3",
    "pyyaml>=6.0.2",
    "click>=8.1.7",
    "rich>=13.9.4",
    "colorama>=0.4.6",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
infrabot = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/infrabot"
Repository = "https://github.com/yourusername/infrabot.git"
Issues = "https://github.com/yourusername/infrabot/issues"

[tool.setuptools]
packages = ["core", "agents", "utils"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
